{{cookiecutter.project_slug}}/{{ cookiecutter.project_slug }}/users/views.py
from braces.views import LoginRequiredMixin

from django.core.urlresolvers import reverse
from django.views.generic import UpdateView

from .models import User


class UserUpdateView(LoginRequiredMixin, UpdateView):

    fields = ['name', ]

    # we already imported User in the view code above, remember?
    model = User

    # send the user back to their own page after a successful update
    def get_success_url(self):
        return reverse('users:update')

    def get_object(self):
        # Only get the User record for the user making the request
        return User.objects.get(pk=self.request.user.pk)
